requirements.txt
amqp==5.2.0
annotated-types==0.6.0
anyio==4.1.0
asgiref==3.5.2
attrs==23.1.0
Automat==22.10.0
beautifulsoup4==4.12.2
billiard==4.2.0
bleach==5.0.1
bs4==0.0.1
celery==5.3.6
certifi==2022.12.7
cffi==1.16.0
channels==3.0.5
charset-normalizer==3.0.1
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cron-descriptor==1.4.0
cryptography==41.0.7
daphne==3.0.2
Django==4.1.6
django-appconf==1.0.5
django-autoslug==1.9.9
django-bootstrap-datepicker-plus==5.0.4
django-braces==1.15.0
django-celery-beat==2.5.0
django-ckeditor==6.5.1
django-ckeditor-5==0.2.3
django-colorfield==0.8.0
django-countries==7.5
django-extra-views==0.8.0
django-filter==1.0.0
django-jazzmin==2.6.0
django-js-asset==2.0.0
django-mathfilters==1.0.0
django-multiselectfield==0.1.12
django-recaptcha==4.0.0
django-render-partial==0.4
django-select2==8.0.0
django-taggit==3.1.0
django-timezone-field==6.1.0
djangorestframework==3.5.3
djongo==1.3.6
dnspython==2.3.0
dongo==0.4.0
future==0.18.3
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
kombu==5.3.4
listview==0.1.0
prompt-toolkit==3.0.41
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
pymongo==3.12.1
pyOpenSSL==23.3.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-telegram-bot==20.7
pytz==2021.3
pyuca==1.2
redis==5.0.1
requests==2.28.2
service-identity==23.1.0
six==1.16.0
sniffio==1.3.0
sorl-thumbnail==12.9.0
soupsieve==2.5
sqlparse==0.2.4
tinycss2==1.1.1
Twisted==23.10.0
txaio==23.1.1
tzdata==2023.3
urllib3==1.26.14
vine==5.1.0
wcwidth==0.2.12
webencodings==0.5.1
zope.interface==6.1
pillow==10.2.0
